Ajout gestion des exceptions

+++ src/main/java/fr/diginamic/rest/controller/AnimalController.java
@@ -1,11 +1,11 @@
 package fr.diginamic.rest.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -18,9 +18,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import fr.diginamic.rest.exceptions.EntityToCreateHasAnIdException;
+import fr.diginamic.rest.exceptions.EntityToUpdateHasNoIdException;
 import fr.diginamic.rest.model.Animal;
-import fr.diginamic.rest.model.Species;
 import fr.diginamic.rest.service.AnimalService;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
 
 @RestController
@@ -35,11 +37,6 @@ public class AnimalController {
 		return animalService.findAll();
 	}
 	
-	@GetMapping("/{id}")
-	public Animal getOneAnimal(@PathVariable("id") Integer id) {
-		return animalService.findById(id);
-	}
-	
 	@GetMapping("/pages")
 	public Page<Animal> findPage(
 			@RequestParam(value = "pageNumber", defaultValue = "0") Integer pageNumber,
@@ -47,10 +44,20 @@ public class AnimalController {
 		return animalService.findPage(PageRequest.of(pageNumber, pageSize));
 	}
 	
+	@GetMapping("/{id}")
+	public Animal getOneAnimal(@PathVariable("id") Integer id) {
+		Optional<Animal> result = Optional.ofNullable(animalService.findById(id));
+		if(!result.isPresent()) {
+			throw new EntityNotFoundException();
+		}
+		return result.get();
+	}
+	
 	@PostMapping
 	public ResponseEntity<Object> createAnimal(@Valid @RequestBody Animal animalItem) {
 		if(animalItem.getId() != null) {
-			return new ResponseEntity<>("Erreur : ID renseigné lors de la création !", HttpStatus.BAD_REQUEST);
+			//return new ResponseEntity<>("Erreur : ID renseigné lors de la création !", HttpStatus.BAD_REQUEST);
+			throw new EntityToCreateHasAnIdException();
 		}
 		this.animalService.create(animalItem);
 		return new ResponseEntity<>("L'animal " + animalItem + " a bien été créé !", HttpStatus.OK);
@@ -59,7 +66,8 @@ public class AnimalController {
 	@PutMapping
 	public ResponseEntity<Object> updateAnimal(@Valid @RequestBody Animal animalItem) {
 		if(animalItem.getId() == null || animalItem.getId() < 0) {
-			return new ResponseEntity<>("Erreur : ID renseigné invalide !", HttpStatus.BAD_REQUEST);
+			//return new ResponseEntity<>("Erreur : ID renseigné invalide !", HttpStatus.BAD_REQUEST);
+			throw new EntityToUpdateHasNoIdException();
 		}
 		this.animalService.create(animalItem);
 		return new ResponseEntity<>("L'animal " + animalItem + " a bien été modifié !", HttpStatus.OK);
@@ -67,6 +75,9 @@ public class AnimalController {
 	
 	@DeleteMapping
 	public void deleteAnimal(@Valid @RequestBody Animal animalItem) {
+		if(animalItem.getId() == null) {
+			throw new EntityNotFoundException();
+		}
 		this.animalService.delete(animalItem);
 	}
