+++ src/main/java/fr/diginamic/rest/controller/PersonController.java
@@ -1,11 +1,11 @@
 package fr.diginamic.rest.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -18,9 +18,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import fr.diginamic.rest.exceptions.EntityToCreateHasAnIdException;
+import fr.diginamic.rest.exceptions.EntityToUpdateHasNoIdException;
 import fr.diginamic.rest.model.Person;
-import fr.diginamic.rest.model.Species;
 import fr.diginamic.rest.service.PersonService;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
 
 @RestController
@@ -44,13 +46,18 @@ public class PersonController {
 	
 	@GetMapping("/{id}")
 	public Person getOnePerson(@PathVariable("id") Integer id) {
-		return personService.findById(id);
+		Optional<Person> result = Optional.ofNullable(personService.findById(id));
+		if(!result.isPresent()) {
+			throw new EntityNotFoundException();
+		}
+		return result.get();
 	}
 	
 	@PostMapping
 	public ResponseEntity<Object> createPerson(@Valid @RequestBody Person personItem) {
 		if(personItem.getId() != null) {
-			return new ResponseEntity<>("Erreur : ID renseigné lors de la création !", HttpStatus.BAD_REQUEST);
+			//return new ResponseEntity<>("Erreur : ID renseigné lors de la création !", HttpStatus.BAD_REQUEST);
+			throw new EntityToCreateHasAnIdException();
 		}
 		this.personService.create(personItem);
 		return new ResponseEntity<>("La personne " + personItem + " a bien été créée !", HttpStatus.OK);
@@ -59,7 +66,8 @@ public class PersonController {
 	@PutMapping
 	public ResponseEntity<Object> updatePerson(@Valid @RequestBody Person personItem) {
 		if(personItem.getId() == null || personItem.getId() < 0) {
-			return new ResponseEntity<>("Erreur : ID renseigné invalide !", HttpStatus.BAD_REQUEST);
+			//return new ResponseEntity<>("Erreur : ID renseigné invalide !", HttpStatus.BAD_REQUEST);
+			throw new EntityToUpdateHasNoIdException();
 		}
 		this.personService.create(personItem);
 		return new ResponseEntity<>("La personne " + personItem + " a bien été modifiée !", HttpStatus.OK);
@@ -67,6 +75,9 @@ public class PersonController {
 	
 	@DeleteMapping
 	public void deletePerson(@Valid @RequestBody Person personItem) {
+		if(personItem.getId() == null) {
+			throw new EntityNotFoundException();
+		}
 		this.personService.delete(personItem);
 	}
 }
