+++ src/main/java/fr/diginamic/rest/controller/SpeciesController.java
@@ -1,11 +1,11 @@
 package fr.diginamic.rest.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -18,8 +18,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import fr.diginamic.rest.exceptions.EntityToCreateHasAnIdException;
+import fr.diginamic.rest.exceptions.EntityToUpdateHasNoIdException;
 import fr.diginamic.rest.model.Species;
 import fr.diginamic.rest.service.SpeciesService;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
 
 @RestController
@@ -43,13 +46,18 @@ public class SpeciesController {
 	
 	@GetMapping("/{id}")
 	public Species getOneSpecies(@PathVariable("id") Integer id) {
+		Optional<Species> result = Optional.ofNullable(speciesService.findById(id));
+		if(!result.isPresent()) {
+			throw new EntityNotFoundException();
+		}
 		return speciesService.findById(id);
 	}
 	
 	@PostMapping
 	public ResponseEntity<Object> createSpecies(@Valid @RequestBody Species speciesItem) {
 		if(speciesItem.getId() != null) {
-			return new ResponseEntity<>("Erreur : ID renseigné lors de la création !", HttpStatus.BAD_REQUEST);
+			//return new ResponseEntity<>("Erreur : ID renseigné lors de la création !", HttpStatus.BAD_REQUEST);
+			throw new EntityToCreateHasAnIdException();
 		}
 		this.speciesService.create(speciesItem);
 		//this.speciesService.create(null);
@@ -59,7 +67,8 @@ public class SpeciesController {
 	@PutMapping
 	public ResponseEntity<Object> updateSpecies(@Valid @RequestBody Species speciesItem) {
 		if(speciesItem.getId() == null || speciesItem.getId() < 0) {
-			return new ResponseEntity<>("Erreur : ID renseigné invalide !", HttpStatus.BAD_REQUEST);
+			//return new ResponseEntity<>("Erreur : ID renseigné invalide !", HttpStatus.BAD_REQUEST);
+			throw new EntityToUpdateHasNoIdException();
 		}
 		this.speciesService.create(speciesItem);
 		return new ResponseEntity<>("L'espèce " + speciesItem + " a bien été modifiée !", HttpStatus.OK);
@@ -67,6 +76,9 @@ public class SpeciesController {
 	
 	@DeleteMapping
 	public void deleteSpecies(@Valid @RequestBody Species speciesItem) {
+		if(speciesItem.getId() == null) {
+			throw new EntityNotFoundException();
+		}
 		this.speciesService.delete(speciesItem);
 	}
